viewer for probabiliy distribution (map) correct per cell place prob (but maybe too fast going to 1-0-0-...)

+++ semantic_mapping/include/semantic_mapping/VisionResult.h
@@ -43,7 +43,7 @@ class VisionResult{
     tf::Pose pose_;
     std::vector<Object> objects_;
     std::vector<PlaceGuess> place_guesses_;
-    std::vector<float> max_dists_;
+    std::vector<double> max_dists_;
 
     //double getSumPlaceProbs() const;
     //double getSumObjectProbs() const;
